schoonheidsfoutje

+++ Notepad/app/src/main/java/com/example/notepad/database/Converters.kt
@@ -14,6 +14,6 @@ class Converters {
 
     @TypeConverter
     fun dateToTimestamp(date: Date?): Long? {
-        return date?.time?.toLong()
+        return date?.time?
     }
 }
